WIP start-vat event
For static vats, initializeSwingset() should push a `start-vat` event onto
the run-queue, before pushing the `deliver(bootstrap)` message.

The `start-vat` event will create a `start-vat` KernelDeliveryObject,
translate it, and deliver it to the vat.

For dynamic vats, processCreateVat() will invoke the same function that the
'start-vat' uses, just after creating the vat and just before deciding
whether the creation succeeded or failed.